requirements.txt
asgiref==3.3.1
boto3==1.17.57
botocore==1.20.57
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
django-cors-headers==3.7.0
django-storages==1.11.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
gunicorn==20.1.0
idna==2.10
jmespath==0.10.0
Pillow==8.1.2
psycopg2==2.8.6
PyJWT==2.0.1
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
requests==2.25.1
s3transfer==0.4.2
six==1.15.0
sqlparse==0.4.1
stripe==2.56.0
urllib3==1.26.4
whitenoise==5.2.0
